Reformat linkedlist copy_random_pointer
  - Remove classes. Now there are 2 versions of the same function
  - Remove tests.
    - Move tests to the tests directory. Since the structure of this
    node is slightly different (adding an atribute) there is a helper
    inner function that will create the nodes in the structure for the 2
    tests that were added.
